+++ src/main/java/com/slcube/shop/business/address/dto/AddressSaveRequestDto.java
@@ -1,13 +1,12 @@
-package com.slcube.shop.business.address;
+package com.slcube.shop.business.address.dto;
 
 import lombok.Getter;
 import lombok.Setter;
 
 @Getter
 @Setter
-public class AddressUpdateRequestDto {
+public class AddressSaveRequestDto {
 
-    private Long addressId;
     private String city;
     private String zipcode;
     private String street;
