+++ src/main/java/com/slcube/shop/business/address/dto/AddressUpdateRequestDto.java
@@ -1,12 +1,13 @@
-package com.slcube.shop.business.address;
+package com.slcube.shop.business.address.dto;
 
 import lombok.Getter;
 import lombok.Setter;
 
 @Getter
 @Setter
-public class AddressSaveRequestDto {
+public class AddressUpdateRequestDto {
 
+    private Long addressId;
     private String city;
     private String zipcode;
     private String street;
