+++ src/main/java/com/slcube/shop/business/cart/domain/Cart.java
@@ -1,7 +1,6 @@
 package com.slcube.shop.business.cart.domain;
 
 import com.slcube.shop.common.domain.BaseEntity;
-import com.slcube.shop.business.cart.CartItem;
 import com.slcube.shop.business.member.Member;
 import lombok.AccessLevel;
 import lombok.Getter;
