+++ src/main/java/com/slcube/shop/business/order/domain/OrderItem.java
@@ -1,7 +1,7 @@
-package com.slcube.shop.business.item;
+package com.slcube.shop.business.order.domain;
 
-import com.slcube.shop.business.category.domain.Category;
 import com.slcube.shop.business.item.domain.Item;
+import com.slcube.shop.business.order.domain.Order;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -11,18 +11,18 @@ import javax.persistence.*;
 @Entity
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @Getter
-public class ItemCategory {
+public class OrderItem {
 
     @Id
     @GeneratedValue
-    @Column(name = "item_category_id")
+    @Column(name = "order_item_id")
     private Long id;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "order_id")
+    private Order order;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "item_id")
     private Item item;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "category_id")
-    private Category category;
 }
