+++ src/main/java/com/slcube/shop/business/order/domain/OrderStatus.java
@@ -1,4 +1,4 @@
-package com.slcube.shop.business.order;
+package com.slcube.shop.business.order.domain;
 
 public enum OrderStatus {
