+++ src/main/java/com/slcube/shop/business/review/domain/Review.java
@@ -2,7 +2,6 @@ package com.slcube.shop.business.review.domain;
 
 import com.slcube.shop.common.domain.BaseEntity;
 import com.slcube.shop.business.member.Member;
-import com.slcube.shop.business.review.ReportedReview;
 import com.slcube.shop.business.item.domain.Item;
 import lombok.AccessLevel;
 import lombok.Getter;
